+++ client/src/common/Header/index.jsx
@@ -9,30 +9,8 @@ import { removeCookie } from "tiny-cookie";
 import { Link } from "react-router-dom";
 import axios from "axios";
 import { getCookie } from "tiny-cookie";
-import { checkerNone, checkerProp } from "../../helpers/helpers";
+import { checkerProp } from "../../helpers/helpers";
 
-// static
-const events = [
-  {
-    title: "Platform Front-end",
-    name: "@ Latte Python ",
-    place: "12 Zybitskaya St., Minsk",
-    time: "22.04.2019 - 16:00",
-  },
-  {
-    title: "Platform Back-end",
-    name: "@ Latte Python ",
-    place: "12 Zybitskaya St., Minsk",
-    time: "22.04.2019 - 16:00",
-  },
-  {
-    title: "CG & Motion Design",
-    name: "@ Latte Python ",
-    place: "12 Zybitskaya St., Minsk",
-    time: "22.04.2019 - 16:00",
-  },
-];
-//
 const getUpcomingEvents = userId => {
   const token = getCookie("token");
   const obj = {
@@ -154,9 +132,9 @@ const userNavigation = (props, userEvents) => {
 };
 
 const Header = props => {
-  const hasId = checkerProp(sessionStorage.getItem("id"));
   const [userEvents, setUserEvents] = useState([]);
   useEffect(() => {
+    const hasId = checkerProp(sessionStorage.getItem("id"));
     const fetchData = async () => {
       if (!hasId) {
         const result = await getUpcomingEvents(sessionStorage.getItem("id"));
