+++ client/src/common/Header/static.js
@@ -1,0 +1,22 @@
+// static
+const events = [
+  {
+    title: "Platform Front-end",
+    name: "@ Latte Python ",
+    place: "12 Zybitskaya St., Minsk",
+    time: "22.04.2019 - 16:00",
+  },
+  {
+    title: "Platform Back-end",
+    name: "@ Latte Python ",
+    place: "12 Zybitskaya St., Minsk",
+    time: "22.04.2019 - 16:00",
+  },
+  {
+    title: "CG & Motion Design",
+    name: "@ Latte Python ",
+    place: "12 Zybitskaya St., Minsk",
+    time: "22.04.2019 - 16:00",
+  },
+];
+//
