+++ client/src/helpers/helpers.js
@@ -1,5 +1,3 @@
-import React from "react";
-
 const checkerProp = prop => {
   if (
     prop === "null" ||
