+++ client/src/pages/HomeAdmin/Teams/index.js
@@ -68,9 +68,9 @@ class DeleteBtn extends Component {
 }
 
 class Teams extends Component {
-  constructor(props) {
-    super(props);
-  }
+  // constructor(props) {
+  //   super(props);
+  // }
 
   state = {
     teams: [],
