+++ client/src/ui/components/Loader.jsx
@@ -1,6 +1,5 @@
 import React from "react";
 import styled from "styled-components";
-import { Link } from "react-router-dom";
 
 export const Loading = () => {
   return (
