+++ client/src/ui/core/home/index.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import styled from "styled-components";
 
 export const TabContainer = styled.div`
@@ -16,8 +15,7 @@ export const Tab = styled.div`
   text-align: center;
   padding: 20px;
   cursor: pointer;
-  border-bottom: ${({ active }) =>
-    active ? "1px solid #0064FF;" : "1px solid #e6e6e6;"};
+  border-bottom: ${({ active }) => (active ? "1px solid #0064FF;" : "1px solid #e6e6e6;")};
 
   &:hover {
     background-color: #f5f5f5;
